+++ src/main/java/com/github/zipcodewilmington/MainApplication.java
@@ -1,11 +1,19 @@
 package com.github.zipcodewilmington;
 
 import com.github.zipcodewilmington.casino.CasinoAccountManager;
+import com.github.zipcodewilmington.casino.games.Klondike;
+
+import java.io.FileNotFoundException;
+
 
 public class MainApplication {
-    public static void main(String[] args) {
-        CasinoAccountManager main = new CasinoAccountManager();
+    public static void main(String[] args) throws FileNotFoundException {
+        Casino main = new Casino();
 
-        main.mainMenu();
+        main.read();
+    main.mainMenu();
+
+
+
     }
 }
