+++ src/main/java/com/github/zipcodewilmington/casino/games/chuckaluck.java
@@ -1,0 +1,4 @@
+package com.github.zipcodewilmington.casino.games;
+
+public class chuckaluck {
+}
